+++ iceoryx2-cal/src/shm_allocator/shm_bump_allocator.rs
@@ -79,7 +79,7 @@ impl ShmAllocator for BumpAllocator {
     }
 
     fn relative_start_address(&self) -> usize {
-        self.allocator.start_address() - self.base_address
+        self.allocator.start_address() as usize - self.base_address
     }
 
     unsafe fn new_uninit(
@@ -179,7 +179,7 @@ impl ShmAllocator for BumpAllocator {
             }
 
             if placement == ContentPlacement::Back {
-                let src = self.allocator.start_address() + offset.offset();
+                let src = self.allocator.start_address() as usize + offset.offset();
                 let dst = src + (new_layout.size() - old_layout.size());
                 unsafe { core::ptr::copy(src as *const u8, dst as *mut u8, old_layout.size()) };
             }
@@ -188,10 +188,10 @@ impl ShmAllocator for BumpAllocator {
         } else {
             match unsafe { self.allocate(new_layout) } {
                 Ok(new_offset) => {
-                    let src = self.allocator.start_address() + offset.offset();
+                    let src = self.allocator.start_address() as usize + offset.offset();
                     match placement {
                         ContentPlacement::Front => {
-                            let dst = self.allocator.start_address() + new_offset.offset();
+                            let dst = self.allocator.start_address() as usize + new_offset.offset();
                             unsafe {
                                 core::ptr::copy_nonoverlapping(
                                     src as *const u8,
@@ -203,7 +203,7 @@ impl ShmAllocator for BumpAllocator {
                             Ok(new_offset)
                         }
                         ContentPlacement::Back => {
-                            let dst = self.allocator.start_address()
+                            let dst = self.allocator.start_address() as usize
                                 + new_offset.offset()
                                 + new_layout.size()
                                 - old_layout.size();
